requirements.txt
crewai>=0.16.0
langchain-groq>=0.0.3
python-dotenv>=1.0.0
crewai-tools>=0.0.9
serper-dev>=0.1.3
langchain>=0.1.0